improve scripting docs
