+++ docs/src/content/addons-events.md
@@ -21,4 +21,18 @@ header with a count of the number of responses seen:
 Below is an addon class that implements stubs for all events. We've added
 annotations to illustrate the argument types for the various events.
 
+### Generic Events
+
 {{< example src="examples/addons/events.py" lang="py" >}}
+
+### HTTP Events
+
+{{< example src="examples/addons/events-http-specific.py" lang="py" >}}
+
+### WebSocket Events
+
+{{< example src="examples/addons/events-websocket-specific.py" lang="py" >}}
+
+### TCP Events
+
+{{< example src="examples/addons/events-tcp-specific.py" lang="py" >}}
