+++ docs/src/themes/mitmproxydocs/layouts/partials/header.html
@@ -20,6 +20,6 @@
     {{ if .RSSLink -}}
     <link href="{{ .RSSLink }}" rel="feed" type="application/rss+xml" title="{{ .Site.Title }}">
     {{- end }}
-    {{ .Hugo.Generator }}
+    {{ hugo.Generator }}
 </head>
 <body>
