+++ examples/addons/events.py
@@ -10,99 +10,6 @@ import mitmproxy.proxy.protocol
 
 
 class Events:
-    # HTTP lifecycle
-    def http_connect(self, flow: mitmproxy.http.HTTPFlow):
-        """
-            An HTTP CONNECT request was received. Setting a non 2xx response on
-            the flow will return the response to the client abort the
-            connection. CONNECT requests and responses do not generate the usual
-            HTTP handler events. CONNECT requests are only valid in regular and
-            upstream proxy modes.
-        """
-
-    def requestheaders(self, flow: mitmproxy.http.HTTPFlow):
-        """
-            HTTP request headers were successfully read. At this point, the body
-            is empty.
-        """
-
-    def request(self, flow: mitmproxy.http.HTTPFlow):
-        """
-            The full HTTP request has been read.
-        """
-
-    def responseheaders(self, flow: mitmproxy.http.HTTPFlow):
-        """
-            HTTP response headers were successfully read. At this point, the body
-            is empty.
-        """
-
-    def response(self, flow: mitmproxy.http.HTTPFlow):
-        """
-            The full HTTP response has been read.
-        """
-
-    def error(self, flow: mitmproxy.http.HTTPFlow):
-        """
-            An HTTP error has occurred, e.g. invalid server responses, or
-            interrupted connections. This is distinct from a valid server HTTP
-            error response, which is simply a response with an HTTP error code.
-        """
-
-    # TCP lifecycle
-    def tcp_start(self, flow: mitmproxy.tcp.TCPFlow):
-        """
-            A TCP connection has started.
-        """
-
-    def tcp_message(self, flow: mitmproxy.tcp.TCPFlow):
-        """
-            A TCP connection has received a message. The most recent message
-            will be flow.messages[-1]. The message is user-modifiable.
-        """
-
-    def tcp_error(self, flow: mitmproxy.tcp.TCPFlow):
-        """
-            A TCP error has occurred.
-        """
-
-    def tcp_end(self, flow: mitmproxy.tcp.TCPFlow):
-        """
-            A TCP connection has ended.
-        """
-
-    # Websocket lifecycle
-    def websocket_handshake(self, flow: mitmproxy.http.HTTPFlow):
-        """
-            Called when a client wants to establish a WebSocket connection. The
-            WebSocket-specific headers can be manipulated to alter the
-            handshake. The flow object is guaranteed to have a non-None request
-            attribute.
-        """
-
-    def websocket_start(self, flow: mitmproxy.websocket.WebSocketFlow):
-        """
-            A websocket connection has commenced.
-        """
-
-    def websocket_message(self, flow: mitmproxy.websocket.WebSocketFlow):
-        """
-            Called when a WebSocket message is received from the client or
-            server. The most recent message will be flow.messages[-1]. The
-            message is user-modifiable. Currently there are two types of
-            messages, corresponding to the BINARY and TEXT frame types.
-        """
-
-    def websocket_error(self, flow: mitmproxy.websocket.WebSocketFlow):
-        """
-            A websocket connection has had an error.
-        """
-
-    def websocket_end(self, flow: mitmproxy.websocket.WebSocketFlow):
-        """
-            A websocket connection has ended.
-        """
-
     # Network lifecycle
     def clientconnect(self, layer: mitmproxy.proxy.protocol.Layer):
         """
